xero/ingestion/code/requirements.txt
flask==2.0.2
werkzeug==2.0.3
apache-beam==2.35.0
google-cloud-storage==1.43.0
google-cloud-secret-manager==2.8.0
requests-oauthlib==1.3.0
structlog==21.5.0
cachetools==4.2.4
gunicorn==20.1.0